# docs/swagger.yaml
basePath: /api/v1
definitions:
  authadmin.LoginAdminRequest:
    properties:
      login:
        example: admin
        maxLength: 50
        type: string
      password:
        example: "12345678"
        minLength: 6
        type: string
    required:
    - login
    - password
    type: object
  authadmin.RegisterAdminRequest:
    properties:
      adminname:
        example: admin
        maxLength: 50
        type: string
      login:
        example: admin
        maxLength: 50
        type: string
      password:
        example: "12345678"
        minLength: 6
        type: string
    required:
    - adminname
    - login
    - password
    type: object
  authemployee.LoginEmployeeRequest:
    properties:
      login:
        example: elogin
        maxLength: 50
        minLength: 4
        type: string
      password:
        example: "12345678"
        minLength: 4
        type: string
    required:
    - login
    - password
    type: object
  authemployee.RegisterEmployeeRequest:
    properties:
      adminID:
        description: Valid    bool      `json:"valid" binding:"required,boolean" example:"true"`
        example: 8f005053-5b95-4a6a-bdcd-7395ee3ed204
        type: string
      login:
        example: elogin
        maxLength: 50
        minLength: 4
        type: string
      password:
        example: "12345678"
        minLength: 4
        type: string
      username:
        example: ename
        maxLength: 50
        type: string
    required:
    - adminID
    - login
    - password
    - username
    type: object
  authmodels.Login2FAUserRequest:
    properties:
      code:
        example: 1234q
        type: string
      sessionAuthID:
        example: cfd9ff5d-cb37-407c-b043-288a482e9239
        type: string
    required:
    - code
    - sessionAuthID
    type: object
  authmodels.Login2FAUserResponse:
    properties:
      access_token:
        type: string
    type: object
  authmodels.LoginUserRequest:
    properties:
      login:
        example: ulogin
        maxLength: 50
        minLength: 4
        type: string
      password:
        example: "12345678"
        minLength: 4
        type: string
    required:
    - login
    - password
    type: object
  authmodels.RegisterUserRequest:
    properties:
      email:
        example: uuser@test.ru
        maxLength: 100
        minLength: 6
        type: string
      login:
        example: ulogin
        maxLength: 50
        minLength: 4
        type: string
      password:
        example: "12345678"
        minLength: 4
        type: string
      subscribe_email:
        example: true
        type: boolean
      username:
        example: uname
        maxLength: 50
        type: string
    required:
    - email
    - login
    - password
    - subscribe_email
    - username
    type: object
...
  jsonreqresp.EmployeeResponse:
    properties:
      adminId:
        example: bb2e8400-e29b-41d4-a716-446655443333
        type: string
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: bb2e8400-e29b-41d4-a716-446655442222
        type: string
      login:
        example: johndoe@example.com
        type: string
      username:
        example: john doe
        type: string
      valid:
        example: true
        type: boolean
    type: object
...
  jsonreqresp.UpdateValidEmployeeRequest:
    properties:
      id:
        example: bb2e8400-e29b-41d4-a716-446655442222
        type: string
      valid:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для системы учета произведений искусств
  title: Музей
  version: "1.0"
paths:
  /admin/employeelist/:
    get:
      consumes:
      - application/json
      description: Возвращает список всех сотрудников
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonreqresp.EmployeeResponse'
            type: array
        "401":
          description: Не авторизован
      security:
      - ApiKeyAuth: []
      summary: Получить всех сотрудников
      tags:
      - Администратор
  /admin/employeelist/change-rights:
    put:
      consumes:
      - application/json
      description: Изменяет поле valid сотрудника
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.UpdateValidEmployeeRequest'
      responses:
        "200":
          description: Успешное обновление
        "404":
          description: Сотрудник не найден
      security:
      - ApiKeyAuth: []
      summary: Изменить права сотрудника
      tags:
      - Администратор
  /admin/employeelist/register-employee:
    post:
      consumes:
      - application/json
      description: Регистрирует нового сотрудника
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authemployee.RegisterEmployeeRequest'
      responses:
        "200":
          description: Сотрудник зарегистрирован
        "400":
          description: Неверные входные параметры
        "401":
          description: Ошибка авторизации
        "409":
          description: Попытка повторной регистрации
      security:
      - ApiKeyAuth: []
      summary: Регистрация сотрудника
      tags:
      - Администратор
  /admin/userlist/:
    get:
      consumes:
      - application/json
      description: Возвращает список всех пользователей
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonreqresp.EmployeeResponse'
            type: array
        "401":
          description: Не авторизован
      security:
      - ApiKeyAuth: []
      summary: Получить всех пользователей
      tags:
      - Администратор
  /auth-admin/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует администратора и возвращает токен доступа
      parameters:
      - description: Учетные данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authadmin.LoginAdminRequest'
      responses:
        "200":
          description: Администратор успешно аутентифицирован
        "400":
          description: Неверные входные параметры
        "401":
          description: Ошибка аутентификации
        "403":
          description: Нет прав доступа
      summary: Вход администратора
      tags:
      - Аутентификация
  /auth-admin/register:
    post:
      consumes:
      - application/json
      description: Register a new admin
      parameters:
      - description: Register credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authadmin.RegisterAdminRequest'
      responses:
        "200":
          description: The admin registered
        "400":
          description: Wrong input parameters
        "401":
          description: Auth error
      summary: Register admin
      tags:
      - Аутентификация
  /auth-employee/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует сотрудника и возвращает токен доступа
      parameters:
      - description: Учетные данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authemployee.LoginEmployeeRequest'
      responses:
        "200":
          description: Сотрудник успешно аутентифицирован
        "400":
          description: Неверные входные параметры
        "401":
          description: Ошибка аутентификации
        "403":
          description: Нет прав доступа
      summary: Вход сотрудника
      tags:
      - Аутентификация
  /auth-employee/register:
    post:
      consumes:
      - application/json
      description: Регистрирует сотрудника
      parameters:
      - description: Учетные данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authemployee.RegisterEmployeeRequest'
      responses:
        "200":
          description: Сотрудник успешно зарегистрирован
        "400":
          description: Неверные входные параметры
        "401":
          description: Ошибка аутентификации
        "403":
          description: Нет прав доступа
      summary: Регистрация сотрудника
      tags:
      - Аутентификация
  /auth-user/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя, присылает код подтверждения на почту
        и возвращает id сессии авторизации
      parameters:
      - description: Учетные данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authmodels.LoginUserRequest'
      responses:
        "200":
          description: Пользователь успешно аутентифицирован и ему отправлен код подтверждения
        "400":
          description: Неверные входные параметры
        "401":
          description: Ошибка аутентификации
      summary: Вход пользователя
      tags:
      - аутентификация
  /auth-user/login-2fa:
    post:
      consumes:
      - application/json
      description: Подтверждает вход пользователя с помощью одноразового кода и выдает
        access token
      parameters:
      - description: Данные для подтверждения входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authmodels.Login2FAUserRequest'
      responses:
        "200":
          description: Успешная аутентификация, возвращает access token
          schema:
            $ref: '#/definitions/authmodels.Login2FAUserResponse'
        "400":
          description: Неверные входные параметры или невалидный session ID
        "401":
          description: Неверный код подтверждения или пользователь не найден
        "500":
          description: Внутренняя ошибка сервера
      summary: Вход с двухфакторной аутентификацией
      tags:
      - аутентификация
  /auth-user/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authmodels.RegisterUserRequest'
      responses:
        "200":
          description: Пользователь зарегистрирован
        "400":
          description: Неверные входные параметры
        "401":
          description: Ошибка аутентификации
        "409":
          description: Попытка повторной регистрации
      summary: Регистрация пользователя
      tags:
      - аутентификация
...
swagger: "2.0"